+++ packages/expo-router/CHANGELOG.md
@@ -11,7 +11,7 @@
 - Fix initial URL when using Expo Go ([#34596](https://github.com/expo/expo/pull/34596) by [@marklawlor](https://github.com/marklawlor))
 - Fix Fast Refresh not detecting file system updates. ([#34509](https://github.com/expo/expo/pull/34509) by [@marklawlor](https://github.com/marklawlor))
 - Fix getPathFromState generating an invalid path for hoisted index routes. ([#34668](https://github.com/expo/expo/pull/34668) by [@marklawlor](https://github.com/marklawlor))
-- Fix `useLocalSearchParams` not updating inside layout routes
+- Fix `useLocalSearchParams` not updating inside layout routes ([#35408](https://github.com/expo/expo/pull/35408) by [@marklawlor](https://github.com/marklawlor))
 
 ### 💡 Others