+++ packages/expo-router/CHANGELOG.md
@@ -4,10 +4,18 @@
 
 ### 🛠 Breaking changes
 
+### 🎉 New features
+
+### 🐛 Bug fixes
+
+### 💡 Others
+
+## 5.0.0-sdk-52-router-patches — 2025-03-28
+
+### 🛠 Breaking changes
+
 - Wrap app in root `<Slot />` navigator ([#35613](https://github.com/expo/expo/pull/35613) by [@marklawlor](https://github.com/marklawlor))
 - Use UNSTABLE_router & add 'dangerouslySingular' prop ([#35595](https://github.com/expo/expo/pull/35595) by [@marklawlor](https://github.com/marklawlor))
-
-### 🎉 New features
 
 ### 🐛 Bug fixes